% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.area.R
\name{SL.area}
\alias{SL.area}
\title{SL.area}
\usage{
SL.area(f, a, b, eps = 1e-06, lim = 10)
}
\arguments{
\item{f}{the integrand function}

\item{a}{the lower limit of integral}

\item{b}{the upper limit of integral}

\item{eps}{accurancy}

\item{lim}{upper bound between partitions}
}
\value{
the result of integrals
}
\description{
an example function calculating integrals by Recursive method
}
\examples{
f <- function(x) 1/x
quad<-SL.area(f,1,5); quad
}
