% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.Newtons.R
\name{SL.Newtons}
\alias{SL.Newtons}
\title{SL.Newtons}
\usage{
SL.Newtons(fun, x, ep = 1e-05, it_max = 100)
}
\arguments{
\item{fun}{functions consisting of equations}

\item{x}{initial variable}

\item{ep}{accuracy}

\item{it_max}{the maximum number of iterations}
}
\value{
solutions of nonlinear equations
}
\description{
an example function solving solutions of nonlinear equations by Newton method
}
\examples{
funs<-function(x){
f<-c(x[1]^2+x[2]^2-5, (x[1]+1)*x[2]-(3*x[1]+1))
J<-matrix(c(2*x[1], 2*x[2], x[2]-3, x[1]+1), nrow=2)
list(f=f,J=J)
}
SL.Newtons(funs, c(0,1))
}
