% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.ruben.test.R
\name{SL.ruben.test}
\alias{SL.ruben.test}
\title{SL.ruben.test}
\usage{
SL.ruben.test(n, r, alpha = 0.05)
}
\arguments{
\item{n}{the number of samples}

\item{r}{correlation coefficient of sample}

\item{alpha}{quantile}
}
\value{
a interval estimation
}
\description{
an example function getting interval estimation in ruben test
}
\examples{
SL.ruben.test(6, 0.8)
SL.ruben.test(25, 0.7)
}
